API

The one function you need to convert URNs to URLs in a markdown source:

CiteDown.rewrite — Function

Rewrite contents of file f with URNs replaced by URLs if they are configured for citable services in the file's YAML header.

rewrite(f)

A publicly exported function so that linkedMarkdownImage can work from a URN and a CiteDown.Settings instance.

CitableImage.linkedMarkdownImage — Function

Compose markdown wrapping IIIF URL for image linked to an image citation tool installation.

linkedMarkdownImage(ict, img, service; ht, caption)

Arguments

ict URL of an instance of the CiteArchitecture ImageCitationTool.

img Cite2Urn for an image.

service IIIFService

ht Height of resulting image in pixels.

caption Caption to embed in resulting linked markdown string.

Compose markdown wrapping IIIF URL for image linked to an image citation tool installation.

linkedMarkdownImage(img, cdsettings)

Arguments

img URN string

cdsettings CiteDown.Settings

Internal structures and functions

CiteDown.Settings — Type

Settings for interpreting URNs in markdown.

CiteDown.pageparts — Function

Split markdown file into YAML header and markdown body. Returns a tuple of 2 strings. If there is no YAML header, the first string will be empty.

pageparts(f)

CiteDown.settings — Function

Create document settings object from a string source.

settings(src)

CiteDown.null_settings — Function

Initialize a Settings instance with empty values.

null_settings()

CiteDown.urlize — Function

Convert URNs in markdown image references to URL values.

urlize(md, cdsettings)